Move some extensions out.
